Changes app into Sinatra mdoular format
