+++ Gemfile
@@ -2,6 +2,7 @@ source "https://rubygems.org"
 ruby "2.1.5"
 
 gem 'sinatra'
+gem 'sinatra-contrib'
 gem 'thin'
 gem 'nokogiri'
 gem 'sass'
