+++ app/public/stylesheets/style.scss
@@ -1,4 +1,4 @@
 p {
-	font-size: 40px;
-	color: red;
+	font-size: 20px;
+	color: blue;
 }
