+++ config.ru
@@ -2,10 +2,10 @@
 # require 'bundler'
 # Bundler.require(:default)
 # require 'sass/plugin/rack'
-require "./app/app"
+require './app'
 
 # use scss for stylesheets
 # Sass::Plugin.options[:style] = :compressed
 # use Sass::Plugin::Rack
 
-run Sinatra::Application
+run Pacman
