+++ spec/parser_spec.rb
@@ -1,3 +1,4 @@
+require 'spec_helper'
 require './app/parser.rb'
 
 RSpec.describe Parser do
